requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0